requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
dataclasses-json>=0.6.0 